+++ cshelve/_factory.py
@@ -10,6 +10,13 @@ def factory(logger: Logger, provider: str) -> ProviderInterface:
     """
     Return the correct module to be used.
     """
+    logger.debug(f"Creating the provider '{provider}'...")
+    res = _factory(logger, provider)
+    logger.debug("Provider created.")
+    return res
+
+
+def _factory(logger: Logger, provider: str):
     if provider == "azure-blob":
         from ._azure_blob_storage import AzureBlobStorage
 
@@ -19,4 +26,5 @@ def factory(logger: Logger, provider: str) -> ProviderInterface:
 
         return InMemory(logger)
 
+    logger.critical("Provider not found.")
     raise UnknownProviderError(f"Provider Interface '{provider}' is not supported.")